pyproject.toml
[tool.poetry]
name = "llm-batch"
version = "0.1.0"
description = "Process documents with an LLM"
authors = ["Yke Rusticus"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^1.10.12"
langchain = "^0.0.251"
logging = "^0.4.9.6"
openai = "^1.0.0"
python-dotenv = "^1.0.0"
pyyaml = "^6.0.1"
typer = "^0.9.0"
pypdf2 = "^3.0.1"
google-cloud-aiplatform = "^1.31.0"
google-generativeai = "^0.1.0"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.25.0"
black = "^23.7.0"
ruff = "^0.0.282"
isort = "^5.12.0"
pytest = "^7.4.0"
mypy = "^1.4.1"

[tool.poetry.scripts]
llm-batch = "llm_batch.cli:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
